refactor(store): move all ui related actions to ui slice
This includes actions related to
    SideBar
    Success Popup message
