+++ src/store/loginPage/userLoginSlice.ts
@@ -44,9 +44,6 @@ export interface IUserInfo {
     };
     loading: boolean;
     error: any;
-    ui: {
-        showLoginSuccessMssg: boolean;
-    };
 }
 
 const initialState: IUserInfo = {
@@ -59,23 +56,12 @@ const initialState: IUserInfo = {
     },
     loading: false,
     error: {},
-    ui: {
-        showLoginSuccessMssg: false,
-    },
 };
 
 const userLoginSlice = createSlice({
     name: 'user',
     initialState,
-    reducers: {
-        SHOW_LOGIN_SUCCESS_MSSG: (state) => {
-            state.ui.showLoginSuccessMssg = true;
-        },
-
-        HIDE_LOGIN_SUCCESS_MSSG: (state) => {
-            state.ui.showLoginSuccessMssg = false;
-        },
-    },
+    reducers: {},
     extraReducers: (builder) => {
         builder.addCase(userLogIn.pending, (state) => {
             state.loading = true;
@@ -101,8 +87,6 @@ const userLoginSlice = createSlice({
         });
     },
 });
-
-export const { SHOW_LOGIN_SUCCESS_MSSG, HIDE_LOGIN_SUCCESS_MSSG } = userLoginSlice.actions;
 
 export default userLoginSlice.reducer;
